refactor: move pool details components to the components folder
